+++ libraries/provider_mapping.rb
@@ -26,7 +26,7 @@ require_relative 'provider_samhain_app_ubuntu'
 require_relative 'provider_samhain_config'
 require_relative 'provider_samhain_service'
 require_relative 'provider_samhain_service_ubuntu_trusty'
-require_relative 'provider_samhain_service_ubuntu_precise'
+require_relative 'provider_samhain_service_ubuntu_legacy'
 
 if Gem::Version.new(Chef::VERSION) < Gem::Version.new('12')
   Chef::Platform.set(resource: :samhain,
@@ -44,6 +44,6 @@ if Gem::Version.new(Chef::VERSION) < Gem::Version.new('12')
                      provider: Chef::Provider::SamhainService::Ubuntu::Trusty)
   Chef::Platform.set(resource: :samhain_service,
                      platform: :ubuntu,
-                     version: '12.04',
-                     provider: Chef::Provider::SamhainService::Ubuntu::Precise)
+                     version: '< 14.04',
+                     provider: Chef::Provider::SamhainService::Ubuntu::Legacy)
 end